+++ package.json
@@ -13,7 +13,7 @@
     "db:studio": "turbo -F @kdx/db studio",
     "db:seed": "turbo -F @kdx/db seed",
     "db:drop": "turbo -F @kdx/db drop",
-    "dev": "turbo watch dev",
+    "dev": "turbo watch --no-daemon dev",
     "dev:kdx": "turbo watch dev -F @kdx/kdx...",
     "dev:email": "turbo watch dev -F @kdx/react-email...",
     "dev:upstash": "turbo watch dev -F @kdx/upstash-dev...",
