+++ packages/api/src/global.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/consistent-type-imports */
 // Use type safe message keys with `next-intl`
 
 //? @kdx/api should only use "api" namespace
