+++ packages/api/src/utils/locales.ts
@@ -1,7 +1,7 @@
 import type { z, ZodSchema } from "zod";
 import { cookies } from "next/headers";
 
-import { defaultLocale } from "@kdx/locales/locales";
+import { defaultLocale } from "@kdx/locales";
 import { getTranslations } from "@kdx/locales/next-intl/server";
 import { createI18nZodErrors } from "@kdx/validators/useI18nZodErrors";
