+++ packages/locales/src/messages/api/en.json
@@ -1,5 +1,17 @@
 {
   "api": {
+    "emails": {
+      "Join invitedByUsername on Kodix": "Join {invitedByUsername} on Kodix",
+      "Join teamName on Kodix": "Join <teamname></teamname> on <site></site>",
+      "Has invited you to the teamName team on Kodix": "Has invited you to the <teamname></teamname> team on <site></site>",
+      "Join the team": "Join the team",
+      "This invitation was intended for": "This invitation was intended for",
+      "This invite was sent from": "This invite was sent from",
+      "located in": "located in",
+      "If you were not expecting this invitation you can ignore this email": "If you were not expecting this invitation, you can ignore this email. If you are concerned about your account's safety, please reply to this email to get in touch with us.",
+      "Hello": "Hello",
+      "or copy and paste this URL into your browser:": "or copy and paste this URL into your browser:"
+    },
     "No Team Found": "No Team Found",
     "Only the team owner can perform this action": "Only the team owner can perform this action",
     "appName is not installed": "{appName} is not installed",
