+++ packages/locales/src/messages/api/pt-BR.json
@@ -1,5 +1,17 @@
 {
   "api": {
+    "emails": {
+      "Has invited you to the teamName team on Kodix": "Convidou você para a equipe {teamName} em {site}",
+      "If you were not expecting this invitation you can ignore this email": "Se você não esperava este convite, pode ignorar este e-mail. \nSe você está preocupado com a segurança da sua conta, responda a este e-mail para entrar em contato conosco.",
+      "Join invitedByUsername on Kodix": "Junte-se a {invitedByUsername} no Kodix",
+      "Join teamName on Kodix": "Junte-se a {teamName} em {site}",
+      "Join the team": "Junte-se à equipe",
+      "This invitation was intended for": "Este convite foi destinado a",
+      "This invite was sent from": "Este convite foi enviado de",
+      "located in": "localizado em",
+      "Hello": "Olá",
+      "or copy and paste this URL into your browser:": "ou copie e cole este URL em seu navegador:"
+    },
     "No Team Found": "Nenhuma equipe encontrada",
     "Admin role not found": "Função de administrador não encontrada",
     "App already installed": "Aplicativo já instalado",
