+++ packages/react-email/src/globals.d.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/consistent-type-imports */
+//? @kdx/react-email should only use "emails" namespace
+type KdxMessages = typeof import("../../locales/src/messages/api/en.json");
+
+type IntlMessages = KdxMessages;
